% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSHM_PlotFunctions.R
\name{barplotfnctvar}
\alias{barplotfnctvar}
\title{Hot spot and hot moment barplot with variance}
\usage{
barplotfnctvar(df, mean, variance, categories, col = "white", main = "",
  xlab = "", ylab = "")
}
\arguments{
\item{df}{the dataframe}

\item{mean}{in quotations is the column name you want to use to create the bars}

\item{variance}{in quotations is the column name you want to use to create the error bars}

\item{categories}{is a list of number or characters.  The categories create the HSHM varaibles.}

\item{col}{default is white, optional color for the barplots}

\item{main}{default is none, creates a header}

\item{xlab}{default is none, creates an x-axis label}

\item{ylab}{default is none, creates a y-axis label}
}
\description{
This function allows you to create 'hot spot' and 'hot moment categories from a dataframe
}
\examples{
x<-c(1,2,3,4,5,6,7,8,9,10,11)
y<-c(12,13,14,15,16,17,18,19,20,21,22)
z<-c(1.3245234,2.2354256,3.34543,4.34634,5.43643,6.3463,7.34643,8.57345,9.65857,10.68454357,11.67986543)
df<-data.frame(x,y,z)
df<-HSHMcategories(df,"y",brkpts=c(11,16.5,23),categories=c("low","middle","high"))
HSHMIndex_df<-HSHMindex(df,"z","HSHM.cat")
baplotfnctvar(HSHMIndex_df,"mean.z","se.z","HSHM.cat",col=c("red","green","blue"))
}
