% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSHMfuncitons.R
\name{HSHMcategories}
\alias{HSHMcategories}
\title{Hot spot and hot moment category}
\usage{
HSHMcategories(df, cutvariable, brkpts, categories, scenario = "cat")
}
\arguments{
\item{df}{the dataframe}

\item{cutvariable}{in quotations is the column name you want to use to break the dataframe into categories}

\item{brkpts}{is a list of numbers to cut the dataset into variables.  The brkpts must be equal the number of categories plus one.  The numbers should bracket the values you want within a category, rather than equal those values.}

\item{categories}{is a list of number or characters.  The categories create the HSHM varaibles.}

\item{scenario}{default is "cat", will create a new category if you would like to create more than one HSHM category from a dataframe (e.g. scenario="spatial" will create a column name "HSHM.spatial)}
}
\description{
This function allows you to create 'hot spot' and 'hot moment categories from a dataframe
}
\examples{
x<-c(1,2,3,4,5,6,7,8,9,10,11)
y<-c(12,13,14,15,16,17,18,19,20,21,22)
df<-data.frame(x,y)
df<-HSHMcategories(df,"y",brkpts=c(11,16.5,23),categories=c("low","middle","high"))
}
