% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSHMfuncitons.R
\name{HSHMindex}
\alias{HSHMindex}
\title{Hot spot and hot moment index}
\usage{
HSHMindex(df1, response, category, dfspatial = df1)
}
\arguments{
\item{df1}{the dataframe}

\item{response}{in quotations, the response variable}

\item{category}{in quotations, the HSHM category variable}

\item{dfspatial}{default is df1, Allows you to use a different spatial dataframe}
}
\description{
This function allows you to calculate a mean, proportion, and index for a HSHM
}
\examples{
x<-c(1,2,3,4,5,6,7,8,9,10,11)
y<-c(12,13,14,15,16,17,18,19,20,21,22)
z<-c(1.3245234,2.2354256,3.34543,4.34634,5.43643,6.3463,7.34643,8.57345,9.65857,10.68454357,11.67986543)
df<-data.frame(x,y,z)
df<-HSHMcategories(df,"y",brkpts=c(11,16.5,23),categories=c("low","middle","high"))
HSHMIndex_df<-HSHMindex(df,"z","HSHM.cat")
}
