% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummStatsFunctions.R
\name{LF}
\alias{LF}
\title{Lower Fence Function}
\usage{
LF(x, n = 1.5)
}
\arguments{
\item{x}{a vector of numbers}

\item{n}{is a constant, default is 1.5, n is often set to 3 to determine extreme outliers assuming a normal distribution.}
}
\description{
This function allows you to calculate the lower fence (Tukey Method) for determining outliers.  Where the UF = Q1 - n(Q3-Q1).
}
\examples{
x<-c(1,2,3,4,5,6,7,8,9,10,11)
UF(x)
}
