% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummStatsFunctions.R
\name{N}
\alias{N}
\title{Number of Observations Function}
\usage{
N(x)
}
\arguments{
\item{x}{a vector}
}
\description{
This function allows you to calculate the number of observation in a set
}
\examples{
x<-c(1,2,3,4,5,6,7,8,9,10,11)
N(x)

x<-c("a","b","c")
N(x)
}
