% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummStatsFunctions.R
\name{se}
\alias{se}
\title{Standard Error Function}
\usage{
se(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\description{
This function allows you to calculate the standard error of a sample
}
\examples{
x<-c(1,2,3,4,5,6,7,8,9,10,11)
se(x)
}
