% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummStatsFunctions.R
\name{UF}
\alias{UF}
\title{Upper Fence Function}
\usage{
UF(x, n = 1.5)
}
\arguments{
\item{x}{a vector of numbers}

\item{n}{is a constant, default is 1.5, n is often set to 3 to determine extreme outliers assuming a normal distribution.}
}
\description{
This function allows you to calculate the upper fence (Tukey Method) for determining outliers.  Where the UF = Q3 + n(Q3-Q1).
}
\examples{
x<-c(1,2,3,4,5,6,7,8,9,10,11)
UF(x)
}
